requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
gradio==4.44.0
sentence-transformers==2.7.0
faiss-cpu==1.12.0
rank-bm25==0.2.2
pymupdf==1.26.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.66.5
matplotlib==3.9.0
ragas==0.1.14
datasets==2.20.0
requests==2.32.3
pydantic==2.8.2
